requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
requests==2.31.0
numpy==1.24.4
torch==2.0.1
faiss-cpu==1.7.4
networkx==3.2.1
pytest==7.4.0
docker==6.1.3